[Fix] Sample tab: Fix to crash that occurred when resizing sample while it was playing. [Fix] Instrument tab: Fix to crash that occurred when multiple windows had instrument tab of the same document open. [Fix] Tuning: Default tuning file should now load correctly (broken in rev. 323). [Ref] S3M load: Check parameters before reading. [Ref] Internal: Changes to get rid of some compiler warnings and disabled some warnings (see Stdafx.h).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@327 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/ScaleEnvPointsDlg.cpp
@@ -9,7 +9,7 @@
 // CScaleEnvPointsDlg dialog
 
 IMPLEMENT_DYNAMIC(CScaleEnvPointsDlg, CDialog)
-CScaleEnvPointsDlg::CScaleEnvPointsDlg(CWnd* pParent, MODINSTRUMENT* pInst, BYTE env)
+CScaleEnvPointsDlg::CScaleEnvPointsDlg(CWnd* pParent, MODINSTRUMENT* pInst, UINT env)
 	: CDialog(CScaleEnvPointsDlg::IDD, pParent),
 	  m_pInstrument(pInst),
 	  m_Env(env)
